edit models, responses and utils

+++ internal/app/domain/balance.go
@@ -1,6 +1,6 @@
 package domain
 
 type Balance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
