+++ internal/app/models/balance.go
@@ -15,7 +15,7 @@ type BalanceRepository interface {
 
 type Withdrawal struct {
 	OrderNum    string
-	Amount      int
+	Amount      float32
 	Status      string
 	ProcessedAt time.Time
 }
